Protected Routing in Backend and Client

+++ client/src/hooks/workouts/useDeleteWorkout.js
@@ -1,10 +1,18 @@
-import { useWorkoutsContext } from "./useWorkoutsContext" 
+import { useWorkoutsContext } from "./useWorkoutsContext";
+import { useAuthContext } from '../auth/useAuthContext';
+
 export const useDeleteWorkout = () => {
-    
+    const { user } = useAuthContext();
     const { dispatch } = useWorkoutsContext()
     const deleteWorkout = async (workout) => {
+        if (!user){
+            return;
+        }
         const response = await fetch('/api/workouts/' + workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {                
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
